2.0

+++ routes/web.php
@@ -2,6 +2,7 @@
 
 use App\Http\Controllers\CategoriasController;
 use App\Http\Controllers\FunkoController;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Route;
 
 /*
@@ -22,13 +23,13 @@ Route::get('/', function () {
 Route::group(['prefix' => 'funkos'], function () {
     Route::get('/', [FunkoController::class, 'index'])->name('funkos.index');
     Route::get('/{id}', [FunkoController::class,'show'])->name('funkos.show');
-    Route::get('/create', [FunkoController::class,'create'])->name('funkos.create');
-    Route::post('/', [FunkoController::class,'store'])->name('funkos.store');
-    Route::get('/{id}/edit', [FunkoController::class, 'edit'])->name('funkos.edit');
-    Route::put('/{id}', [FunkoController::class, 'update'])->name('funkos.update');
-    Route::delete('/{id}', [FunkoController::class, 'destroy'])->name('funkos.destroy');
-    Route::get('/{id}/edit-image', [FunkoController::class, 'editImage'])->name('funkos.editImage');
-    Route::patch('/{id}/edit-image', [FunkoController::class, 'updateImage'])->name('funkos.updateImage');
+    Route::get('/create', [FunkoController::class,'create'])->name('funkos.create')->middleware(['auth', 'admin']);
+    Route::post('/', [FunkoController::class,'store'])->name('funkos.store')->middleware(['auth', 'admin']);
+    Route::get('/{id}/edit', [FunkoController::class, 'edit'])->name('funkos.edit')->middleware(['auth', 'admin']);
+    Route::put('/{id}', [FunkoController::class, 'update'])->name('funkos.update')->middleware(['auth', 'admin']);
+    Route::delete('/{id}', [FunkoController::class, 'destroy'])->name('funkos.destroy')->middleware(['auth', 'admin']);
+    Route::get('/{id}/edit-image', [FunkoController::class, 'editImage'])->name('funkos.editImage')->middleware(['auth', 'admin']);
+    Route::patch('/{id}/edit-image', [FunkoController::class, 'updateImage'])->name('funkos.updateImage')->middleware(['auth', 'admin']);
 });
 
 
@@ -36,11 +37,15 @@ Route::group(['prefix' => 'funkos'], function () {
 
 
 Route::group(['prefix' => 'categorias'], function () {
-    Route::get('/', [CategoriasController::class, 'index'])->name('categorias.index');
-    Route::get('/{id}', [CategoriasController::class,'show'])->name('categorias.show');
-    Route::get('/create', [CategoriasController::class,'create'])->name('categorias.create');
-    Route::post('/', [CategoriasController::class,'store'])->name('categorias.store');
-    Route::get('/{id}/edit', [CategoriasController::class, 'edit'])->name('categorias.edit');
-    Route::put('/{id}', [CategoriasController::class, 'update'])->name('categorias.update');
-    Route::delete('/{id}', [CategoriasController::class, 'destroy'])->name('categorias.destroy');
+    Route::get('/', [CategoriasController::class, 'index'])->name('categorias.index')->middleware(['auth', 'user']);
+    Route::get('/{id}', [CategoriasController::class,'show'])->name('categorias.show')->middleware(['auth', 'user']);
+    Route::get('/create', [CategoriasController::class,'create'])->name('categorias.create')->middleware(['auth', 'admin']);
+    Route::post('/', [CategoriasController::class,'store'])->name('categorias.store')->middleware(['auth', 'admin']);
+    Route::get('/{id}/edit', [CategoriasController::class, 'edit'])->name('categorias.edit')->middleware(['auth', 'admin']);
+    Route::put('/{id}', [CategoriasController::class, 'update'])->name('categorias.update')->middleware(['auth', 'admin']);
+    Route::delete('/{id}', [CategoriasController::class, 'destroy'])->name('categorias.destroy')->middleware(['auth', 'admin']);
 });
+
+Auth::routes();
+
+Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
